separating out main python script from what user calls
