+++ wisdem/__init__.py
@@ -1,1 +1,1 @@
-from wisdem.glue_code.runWISDEM import run_wisdem
+
