+++ wisdem/inputs/geometry_defaults.yaml
