+++ wisdem/inputs/geometry_schema.yaml
