--- wisdem/yaml/__init__.py
@@ -1,3 +1,0 @@
-from wisdem.yaml.validation import *
-#from wisdem.yaml import gui
-
